refactor : 컨트롤러 서비스 로직으로 리팩토링
