+++ dashboard/src/main/java/com/iroom/dashboard/controller/DashBoardController.java
@@ -25,9 +25,7 @@ public class DashBoardController {
 	private final DashBoardService dashBoardService;
 	private final ChatService chatService;
 	private final PdfService pdfService;
-	private final RestTemplate restTemplate = new RestTemplate();
-	private final EmbeddingService embeddingService;
-	private final String QDRANT_SEARCH_URL = "http://localhost:6333/collections/safety_db/points/search";
+
 
 	//대시보드 조회하기
 	@GetMapping(value = "/{metricType}", produces = "application/json;charset=UTF-8")
@@ -41,52 +39,15 @@ public class DashBoardController {
 		value = "/report"
 	)
 	public ResponseEntity<byte[]> exportReport(@RequestBody ReportRequest reportRequest) throws Exception {
-		int missingPPECount = 40;
-		int dangerZoneAccessCount = 20;
-		int healthAlertCount = 10;
+
 		// 1. 질의 프롬프트 생성
 		String userPrompt = String.format(
 			"오늘 보호구 미착용 " + reportRequest.missingPpeCnt() + "건, 위험지역 접근 " + reportRequest.dangerZoneAccessCnt()
-				+ "건, 건강 이상 알림 " + reportRequest.healthAlertCnt() + "건이 있었습니다. 이에 따른 안전 보고서를 작성해주세요.",
-			missingPPECount, dangerZoneAccessCount, healthAlertCount);
-
-		// 2. Qdrant에 유사 문단 질의 (임의 벡터 사용 중이라 실제 의미 없음)
-		//float[] dummyVector = new float[]{0f, 0f, 0f, 0f, 0f};
-		float[] embeddedVector = embeddingService.embed(userPrompt);
-
-		Map<String, Object> qdrantRequest = new HashMap<>();
-		//qdrantRequest.put("vector", dummyVector);
-		qdrantRequest.put("vector", embeddedVector);
-		qdrantRequest.put("limit", 5);
-		qdrantRequest.put("with_payload", true);
-
-		HttpHeaders headers = new HttpHeaders();
-		headers.setContentType(MediaType.APPLICATION_JSON);
-		HttpEntity<Map<String, Object>> request = new HttpEntity<>(qdrantRequest, headers);
-
-		ResponseEntity<Map> qdrantResponse = restTemplate.postForEntity(
-			QDRANT_SEARCH_URL,
-			request,
-			Map.class
+				+ "건, 건강 이상 알림 " + reportRequest.healthAlertCnt() + "건이 있었습니다. 이에 따른 안전 보고서를 작성해주세요."
 		);
 
-		// 3. Qdrant 응답에서 관련 텍스트 추출
-		StringBuilder contextBuilder = new StringBuilder();
-		try {
-			var resultList = (Iterable<Map<String, Object>>)qdrantResponse.getBody().get("result");
-			for (Map<String, Object> item : resultList) {
-				Map<String, Object> payload = (Map<String, Object>)item.get("payload");
-				if (payload != null && payload.containsKey("content")) {
-					contextBuilder.append(payload.get("content")).append("\n\n");
-				}
-			}
-		} catch (Exception e) {
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-				.body(("Qdrant 응답 처리 실패: " + e.getMessage()).getBytes(StandardCharsets.UTF_8));
 
-		}
-
-		String context = contextBuilder.toString();
+		String context = dashBoardService.getContext(reportRequest,userPrompt);
 		String finalPrompt = context + "\n" + userPrompt;
 
 		// 4. GPT에 최종 질의
